pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whoop-connecter"
version = "0.1.0"
description = "WHOOP MCP server and CLI for OpenClaw Coach"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.scripts]
whoop = "cli.main:app"
whoop-mcp = "mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["whoop", "mcp_server", "cli"]
